Refactor server in it's own module
